fix: Optimize main.dart
