+++ lib/presentation/app.dart
@@ -1,0 +1,75 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:dienstplan/core/l10n/app_localizations.dart';
+import 'package:dienstplan/presentation/screens/app_initializer_widget.dart';
+import 'package:dienstplan/core/services/controller_service.dart';
+import 'package:dienstplan/data/services/language_service.dart';
+import 'package:get_it/get_it.dart';
+
+class MyApp extends StatefulWidget {
+  const MyApp({super.key});
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  LanguageService? _languageService;
+  final RouteObserver<ModalRoute<void>> _routeObserver =
+      RouteObserver<ModalRoute<void>>();
+
+  @override
+  void initState() {
+    super.initState();
+    _initializeControllers();
+  }
+
+  void _initializeControllers() {
+    // Initialize controllers in background without blocking the UI
+    ControllerService.initializeControllers();
+
+    GetIt.instance.getAsync<LanguageService>().then((service) {
+      _languageService = service;
+      // Update UI when language service is ready
+      if (mounted) {
+        setState(() {});
+      }
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // Use a fallback locale if language service is not ready yet
+    final currentLocale = _languageService?.currentLocale ?? const Locale('de');
+
+    return MaterialApp(
+      title: 'Dienstplan',
+      navigatorObservers: [_routeObserver],
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: const Color(0xFF005B8C),
+          primary: const Color(0xFF005B8C),
+        ),
+        useMaterial3: true,
+        appBarTheme: const AppBarTheme(
+          backgroundColor: Color(0xFF005B8C),
+          titleTextStyle: TextStyle(
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+            fontSize: 20,
+          ),
+          iconTheme: IconThemeData(color: Colors.white),
+        ),
+      ),
+      localizationsDelegates: const [
+        AppLocalizations.delegate,
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+      ],
+      supportedLocales: AppLocalizations.supportedLocales,
+      locale: currentLocale,
+      home: AppInitializerWidget(routeObserver: _routeObserver),
+    );
+  }
+}
